pyproject.toml
[project]
name = "classic-car-insights"
version = "0.1.0"
description = "Classic Car Insights App"
authors = [
    {name = "Jonathan Hamler",email = "jkhamler@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.125.0,<0.126.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"